+++ changelog.md
@@ -1,3 +1,1 @@
-* Added versioned config.
-* Change name methods.
-* Update depends.
+* Remove versioned config.